purchase order update: done

+++ backend/controllers/PurchaseOrderController.js
@@ -3,6 +3,7 @@ const PurchaseOrder = require("../models/purchase-order");
 const Supplier = require("../models/supplier");
 const Product = require("../models/product");
 const ProductOrder = require("../models/product-order");
+const { ProductOrderedStatus } = require("shared/enums/purchase-order");
 
 module.exports = {
   all: async (req, res) => {
@@ -37,14 +38,20 @@ module.exports = {
       data.products.forEach(async (product) => {
         await purchaseOrder.addProduct(product.product_id, {
           through: {
-            quantity: product.quantity,
+            cost: product.cost,
             amount: product.amount,
+            quantity: product.quantity,
+            description: product.description,
+            status: product.status ? product.status : ProductOrderedStatus.OPEN,
+            quantity_received: product.quantity_received
+              ? product.quantity_received
+              : 0,
           },
         });
       });
       res.sendResponse({}, "Successfully created!");
     } catch (e) {
-      res.sendError(e, "Something wen't wrong! => " + e.messge);
+      res.sendError({ ...e }, "Something wen't wrong! => " + e.messge);
     }
   },
 
@@ -57,23 +64,64 @@ module.exports = {
             id: req.params.id,
           },
         });
-      }
 
-      if (data.products) {
-        await ProductOrder.destroy({
-          where: {
-            product_id: req.params.id,
-          },
-        });
+        // update related products
+        if (data.products) {
+          const products = await ProductOrder.findAll({
+            where: { order_id: req.params.id },
+            attributes: ["id", "product_id"],
+          });
 
-        for (let product in data.products) {
-          console.log(product);
+          const currentProductIds = products.map((p) => p.product_id);
+          console.log(currentProductIds, data.products);
+
+          await Promise.all([
+            // determine the products in data.products needs to be created
+            ...data.products
+              .filter((p) => !currentProductIds.includes(p.product_id))
+              .map((p) =>
+                ProductOrder.create({ ...p, order_id: req.params.id })
+              ),
+
+            // determine the products weither they it will be updated or deleted
+            ...products.map((product) => {
+              // TODO: Need to check if the data has changes to update
+              const dataProductIds = data.products.map((p) => p.product_id);
+
+              if (dataProductIds.includes(product.product_id)) {
+                const toUpdateFields = data.products.find(
+                  (p) => p.product_id == product.product_id
+                );
+
+                return ProductOrder.update(toUpdateFields, {
+                  where: {
+                    id: product.id,
+                  },
+                });
+              } else {
+                return ProductOrder.destroy({
+                  where: {
+                    id: product.id,
+                  },
+                });
+              }
+            }),
+          ]);
+        }
+
+        // update purchase order address
+        if (data.address) {
+          await Address.update(data.address, {
+            where: {
+              order_id: req.params.id,
+            },
+          });
         }
       }
 
       res.sendResponse({}, "Successfully updated!", 200);
     } catch (e) {
-      res.sendError({}, "Something wen't wrong!", 400);
+      res.sendError({ ...e }, "Something wen't wrong! => " + e.message, 400);
     }
   },
